pyproject.toml
[tool.poetry]
name = "ecommerce-integration"
version = "0.1.0"
description = "An e-commerce integration system for multiple platforms"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
pydantic = "^1.8.2"
python-dotenv = "^0.19.0"
asyncio = "^3.4.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-asyncio = "^0.15.1"
black = "^21.9b0"
isort = "^5.9.3"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
ignore_missing_imports = true
strict_optional = true